pyproject.toml
[tool.poetry]
name = "exchange-data-collector"
version = "0.1.0"
description = ""
authors = ["Shadow Insights"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
websockets = "^12"
prometheus-client = "^0.17.1"
alembic = "^1.12.0"
sqlalchemy = "^2.0.21"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
discord-webhook = "^1.3.0"
regex = "^2023.10.3"
greenlet = "^3.0.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.28.0"
types-regex = "^2023.10.3.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pyproject = "^1.3.1"
pytest-asyncio = "^0.21.1"
mypy = "1.0.1"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.mypy]
mypy_path = "stub"
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
files = ["app", "tests"]
follow_imports = "silent"
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]
asyncio_mode = "auto"
log_level = "info"